mrfixit epic update rebase kernel against latest rockchip bps commit update to 4.4.202 enable virtualization for issue #4 enable megaraid for issue #3 complete leakage settings for cpu overclocks add edp panel timing for boe nv140fhmn49 in pinebook pro a bunch of wifi drivers added/fixed/updated add and enable a bunch of pcmcia drivers additional mainline backports additional pcie drivers and patches fixes for some dvb drivers

+++ drivers/net/wireless/rockchip_wlan/rtl8723au/include/usb_osintf.h
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Copyright(c) 2007 - 2011 Realtek Corporation. All rights reserved.
- *                                        
+ *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of version 2 of the GNU General Public License as
  * published by the Free Software Foundation.
@@ -25,14 +25,13 @@
 #include <drv_types.h>
 #include <usb_vendor_req.h>
 
-#define USBD_HALTED(Status) ((ULONG)(Status) >> 30 == 3)
+#define USBD_HALTED(Status) ((u32)(Status) >> 30 == 3)
 
 
-//uint usb_dvobj_init(_adapter * adapter);
-//void usb_dvobj_deinit(_adapter * adapter);
+//uint usb_dvobj_init(struct rtw_adapter * adapter);
+//void usb_dvobj_deinit(struct rtw_adapter * adapter);
 
 u8 usbvendorrequest(struct dvobj_priv *pdvobjpriv, RT_USB_BREQUEST brequest, RT_USB_WVALUE wvalue, u8 windex, void* data, u8 datalen, u8 isdirectionin);
 
 
 #endif
-
